package/kodi: add Kconfig options to remove default skins
The skins estuary and estouchy are new in Kodi 17-Krypton and replace
the skin confluence, used until Kodi 16-Jarvis.

This patch allows to disable one or both of the new skins, it also takes
care that at least one skin needs to be enabled. With this patch Kodi
can be used only with skin confluence.

Signed-off-by: Bernd Kuhls <bernd.kuhls@t-online.de>