package/eudev: allow building only libudev
We need to move the kconfig symbol for the daemon outside of the eudev
conditional block, or we would, later, ahve circular dependencies, like
so (here, with libinput converted to use libudev):

    package/udev/Config.in:1:error: recursive dependency detected!
    package/udev/Config.in:1:       symbol BR2_PACKAGE_HAS_UDEV is selected by BR2_PACKAGE_EUDEV_DAEMON
    package/eudev/Config.in:1:      symbol BR2_PACKAGE_EUDEV_DAEMON is selected by BR2_PACKAGE_EUDEV
    package/eudev/Config.in:11:     symbol BR2_PACKAGE_EUDEV is selected by BR2_PACKAGE_LIBUDEV
    package/libudev/Config.in:1:    symbol BR2_PACKAGE_LIBUDEV is selected by BR2_PACKAGE_LIBINPUT
    package/libinput/Config.in:1:   symbol BR2_PACKAGE_LIBINPUT is selected by BR2_PACKAGE_WESTON
    package/weston/Config.in:6:     symbol BR2_PACKAGE_WESTON depends on BR2_PACKAGE_HAS_UDEV

Although the chain is short, it is rather difficult to interpret.
Basically, what this means is that, by way of successive 'select' and
'depends on', we get a circular dependency. In fact, kconfig is wrong
here, since there is *no* way we could trigger this loop by enabling one
of the incriminated symbol: its dependency tracking is limited to
looking at the symbols, not whether they are negated or not (i.e. in the
chanin above, a symbol is negated, and is only valid when another is
false, thus avoiding the loop). But kconfig has no way to know that...

So, we trick it and break the loop, by moving the kconfig symbol for the
eudev daemon outside of the eudev if-block and making it promptless.

Still, for the convenience of the user, we keep a dummy udev daemon
prompt in that if-block, so that he knows the daemon is enabled.

Fix the comments on dependencies, since they are also hard-dependencies
of eudev itself.

For wchar, the reason is that configure.ac wants to check for a
C99-compliant compiler, and autoconf always emits test-code that
contains wchar_t variables, even if that's not required at all by eudev.
This is not solved by an autoreconf either, obviously.

For !static, libudev (and thus eudev) includes dlfcn.h, which only
exists in toolchains that have dynamic library support. However, eudev
does not use anything provided by dlfcn.h; removing this dependency can
be done in a follow-up change.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>