package/libudev: make it selectable
Usually, it is not possible for a pacakge to select a virtual package it
depends on, and expects a provider to suddenly be available.

However, for libudev, this is slightly simpler: there will be only two
providers ever: eudev and systemd. Both are incompatible one with the
other, systemd is only available if chosen as an init system.

So, if systemd is not enabled (as an init system), then we can
forcefully enable eudev as the provider for libudev.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>